Create README.md
Welcome to my Virtualization Skills Showcase repository! This repository demonstrates my proficiency in working with KVM (Kernel-based Virtual Machine) and Virtual Machine Manager (VMM) through a series of tasks designed to highlight my skills in creating, managing, and configuring virtual machines.

1. Installing KVM on a bare metal Ubuntu Machine
2. Create a VM using VMM (Virtual Machine Manager) GUI
3. Create a KVM-based VM using virt-install CLI
4. Shared Folder Between Host OS and Guest OS using CLI
5. Connect the Guest OS with an Android Phone
6. Clone a VM using GUI and KVM-based Command
7. Add Two Hard Disks in a New Cloned Virtual Machine using GUI
8. Add Two Hard Disks in a New Cloned Virtual Machine using KVM-based Command

How to Use This Repository

Documentation: Start by reading the detailed documentation provided for each task. The documentation includes commands, screenshots, and descriptions to help you understand the steps involved.

Scripts and Commands: Use the provided scripts and commands to replicate the setup and configurations on your own system.

Learn and Experiment: Follow the guides to learn about KVM and VMM, and experiment with the commands and configurations to enhance your understanding of virtualization.